ARM: dts: Set VAUX1 and VAUX4 to 3.0V and 1.8V respectively
The development kit schematic expects VAUX1 to be 3.0V.  Most users use the development kit as a reference.
The development kit schematic expects VAUX4 to be 1.8V.  VAUX4 powers VDDS_CSI2 on processor.  If the voltage is too high it could damage the processor.
If it's too low, it won't work.

Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: 2d11961f3e554ca7f90ce3ee2ed42421edd1e776 ]