Merge branch 'mpls-multipath-improvements'
Robert Shearman says:

====================
mpls: mulipath improvements

Two improvements to the recently added mpls multipath support. The
first is a fix for missing initialisation the nexthop address length
for the v4 and v6 explicit null label routes, and the second is to
reduce the amount of memory used by mpls routes by changing the way
the via addresses are stored.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 5b9e3bd5b2f807f66a959cc079c495915f032c67 ]