Merge branch 'tipc-cleanups-improvements'
Jon Maloy says:

====================
tipc: some cleanups and improvements

This series mostly contains cleanups and cosmetic code changes.
The only real functional change is in #4 and #5, where we change the
locking structure for nodes and links in order to permit full
concurrency between links working in parallel on different interfaces.
Since the groundwork for this has been done in previous commit series,
this change constitutes only the final, small step to achieve that goal.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 6b99c6d558bbfaba0292db1996199325a74c7d02 ]