Merge branch 'ipmr-nl'
Nikolay Aleksandrov says:

====================
net: ipmr: more cleanups and mfc netlink support

This set continues with the minor cleanups in the first 6 patches and
patch 7 adds the first new feature - MFC manipulation via netlink. It
registers NEWROUTE/DELROUTE for that purpose and uses the same semantics
as the already present netlink dump. The only new attribute that is used
is RTA_PREFSRC to denote an MFC_PROXY entry. Currently the table must
exist before adding an entry, and new tables can be created only via
setsockopt, but that will be changed in the future.
This set was tested with modified iproute2 which supports NEWROUTE/DELROUTE
for RTNL_FAMILY_IPMR.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 75c4a57ca9ee747f1cbb24591d92a41b2eddff52 ]