irqchip/sunxi-nmi: Rename binding doc filename to allwinner,sunxi-nmi.txt
The NMI controller is found in all Allwinner multi-core SoCs. It is not
limited to sun[67]i, nor is it always found in the "system controller"
block. On sun[68]i, it is in the RTC block, while on sun9i, it is in the
PRCM block.

Drop these 2 specific bits from the binding doc filename.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Acked-by: Maxime Ripard <maxime.ripard@free-electrons.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Link: https://lkml.kernel.org/r/1449130813-22400-2-git-send-email-wens@csie.org
Signed-off-by: Jason Cooper <jason@lakedaemon.net>

[ upstream commit: 57e7b08263d046ce0bbdf36adb20a268a80bf60a ]