Merge branch 'mvneta-ethtool-autoneg'
Stas Sergeev says:

====================
mvneta: implement ethtool autonegotiation control

These 2 patches add an ability to control the
autonegotiation via ethtool. For example:

ethtool -s eth0 autoneg off
ethtool -s eth0 autoneg on

This is needed if you want to connect the mvneta's MII
to different switches or PHYs: the ones the do support
the in-band status, and the ones that do not.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: b69e3c6f71a53fb46bba588e651d943785398220 ]